requirements.txt
flask>=3.0.0
cryptography>=42.0.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
pytest>=8.0.0
Werkzeug>=3.0.0
